refactor(db): move reaction store ownership
